// src/app/Models/cts-api.application.ts
export interface Application {
    achCode: string; // number;
    applicationId: number;
    billingCycle: string;
    collateralCode: string;
    collateralCodeDesc: string;
    dealerId: string;
    docProfileId: number;
    docProfileFooter: string;
    docProfileName: string;
    eotOption: string;
    eotOptionDesc?: any;
    gracePeriodDays: number;
    indirectBilling: string; // number;
    indirectBillingDesc: string;
    insuranceStatus: string;
    insuranceStatusDesc: string;
    invoiceCode: string;
    invoiceCodeDesc: string;
    invoiceDesc: string;
    lateChargeExempt: boolean;
    lateChargeMax?: any;
    lateChargeMin: number;
    lateChargePctOfPmt: number;
    leadDays: number;
    leaseType: string;
    leaseTypeDesc: string;
    lesseeAddress1: string;
    lesseeAddress2?: any;
    lesseeAttnName: string;
    lesseeCity: string;
    lesseeCountry: string;
    lesseeEmail: string;
    lesseeFax: string;
    lesseeName: string;
    lesseePhone: string;
    lesseeState: string;
    lesseeZip: string;
    masterAgreementNumber?: any;
    personalGuarantors: any[];
    privateLabel: boolean;
    programType: string;
    programTypeDesc: string;
    purchaseOption: string;
    purchaseOptionDesc: string;
    purchaseOrder: string;
    purchaseOrderInfoLeaseFlag?: any;
    relationshipCode?: any;
    relationshipCodeDesc?: any;
    renewalTerm: string; // number;
    renewalTermDesc: string;
    securExcludeReason: string;
    securExcludeReasonDesc?: any;
    securitizationCode: string;
    securitizationCodeDesc: string;
    signerName?: any;
    signerTitle?: any;
    signerUserNumber?: any;
    term: number;
    totalFinanceAmount: number;
    variablePayment: boolean;
    vendorContractId?: any;
}